components/image_analyzer/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make freeze
#
attrs==21.2.0
boto3==1.18.31
botocore==1.21.31
cycler==0.10.0
decorator==5.0.9
imageio==2.9.0
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
opencv-python==4.5.3.56
pika==1.2.0
pillow==8.3.1
py==1.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pywavelets==1.1.1
retry==0.9.2
s3transfer==0.5.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
smart-open[s3]==5.2.0
tifffile==2021.8.8
urllib3==1.26.6
yolo-v4==0.5
